+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+  - remove lifecycle and connectivity listeners
+
+  - add debounce on start\stop methods
+
 ## 0.0.2
 
   - use onStateChange() callback